refactor: remove unused Go API clients
