+++ go/api/client/clientset.go
@@ -4,16 +4,9 @@ package client
 type ClientSet struct {
 	baseClient *BaseClient
 
-	Health              Health
-	Version             Version
-	ModelConfig         ModelConfigInterface
-	Tool                Tool
-	ToolServer          ToolServer
-	ModelProviderConfig ModelProviderConfig
-	Model               Model
-	Namespace           Namespace
-	AgentInstance       *AgentInstanceClient
-	A2A                 *A2AClient
+	Version       Version
+	AgentInstance *AgentInstanceClient
+	A2A           *A2AClient
 }
 
 // New creates a new KAgent client set
@@ -21,17 +14,10 @@ func New(baseURL string, options ...ClientOption) *ClientSet {
 	baseClient := NewBaseClient(baseURL, options...)
 
 	return &ClientSet{
-		baseClient:          baseClient,
-		Health:              NewHealthClient(baseClient),
-		Version:             NewVersionClient(baseClient),
-		ModelConfig:         NewModelConfigClient(baseClient),
-		Tool:                NewToolClient(baseClient),
-		ToolServer:          NewToolServerClient(baseClient),
-		ModelProviderConfig: NewModelProviderConfigClient(baseClient),
-		Model:               NewModelClient(baseClient),
-		Namespace:           NewNamespaceClient(baseClient),
-		AgentInstance:       NewAgentInstanceClient(baseClient),
-		A2A:                 NewA2AClient(baseClient),
+		baseClient:    baseClient,
+		Version:       NewVersionClient(baseClient),
+		AgentInstance: NewAgentInstanceClient(baseClient),
+		A2A:           NewA2AClient(baseClient),
 	}
 }
